Face detection bounds are reports as precentages.

+++ app/model/DetectedFace.scala
@@ -1,6 +1,6 @@
 package model
 
-case class Point(x: Int, y: Int)
+case class Point(x: Double, y: Double)
 
 case class Bounds(topLeft: Point, bottomRight: Point)
